+++ modules/advertise/database/factories/AdvertisementFactory.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Database\Factories\Advertise;
+namespace Modules\Advertise\Database\Factories;
 
 use App\Models\Geo\City;
 use App\Models\User;
