+++ modules/advertise/src/Models/Gallery.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Modules\Advertise\Models;
 
 use App\Models\Scopes\LatestScope;
-use Database\Factories\Advertise\GalleryFactory;
 use Illuminate\Database\Eloquent\Attributes\ObservedBy;
 use Illuminate\Database\Eloquent\Attributes\ScopedBy;
 use Illuminate\Database\Eloquent\Attributes\UseFactory;
@@ -15,6 +14,7 @@ use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Prunable;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\SoftDeletes;
+use Modules\Advertise\Database\Factories\GalleryFactory;
 use Modules\Advertise\Observers\GalleryObserver;
 
 #[ScopedBy([LatestScope::class])]
